Evaluation logic extraction
