+++ src/commonMain/kotlin/operations/array/unwrap/EvaluatingUnwrapStrategy.kt
@@ -1,17 +1,14 @@
 package operations.array.unwrap
 
-import JsonLogicEngine
-import kotlin.collections.Map
+import LogicEvaluator
 
-internal interface EvaluatingUnwrapStrategy {
-    fun evaluate(expression: Map<String, Any?>, data: Any?) = JsonLogicEngine.instance.evaluate(expression, data)
-
+internal interface EvaluatingUnwrapStrategy : LogicEvaluator {
     fun unwrapOperationData(expression: List<Any?>, data: Any?) =
         (expression.firstOrNull().unwrapOperationData(data) as? List<*>)
 
     private fun Any?.unwrapOperationData(data: Any?): Any? = when {
         this is List<*> -> map { it.unwrapOperationData(data) }
-        isExpression(this) -> evaluate(this as Map<String, Any?>, data)
+        isExpression(this) -> evaluateLogic(this as Map<String, Any?>, data)
         else -> this
     }
